02.Project.py

+++ 02.Project.py
@@ -11,7 +11,7 @@
 A=float(input("Enter side 1: "))     #our sides need to be in float form             
 B=float(input("Enter side 2: "))
 C=float(input("Enter side 3: "))
-P= A + B + C                           #Calculate Perimeter is Adding all three sides 
-S= P/2                                 #Calculate Semiperimeter by diving the perimeter by 2 
+P= float(A + B + C)                           #Calculate Perimeter is Adding all three sides 
+S= float(P/2)                                 #Calculate Semiperimeter by diving the perimeter by 2 
 area= float(S*(S-A)*(S-B)*(S-C))**0.5   #Calculate area by multiplying the s by the out comes of s minus the three sides then mulitplaying it by 0.5 #in this context 0.5 is equalivelnt to taking a squre root
 print ("area:",area)                    #the area must be float form
